Añadido mi tema

+++ protected/controllers/UnidadController.php
@@ -110,11 +110,13 @@ class UnidadController extends Controller
 	 */
 	public function actionDelete($id)
 	{
+        if ($id==1){
 		$this->loadModel($id)->delete();
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		if(!isset($_GET['ajax']))
 			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+        }
 	}
 
 	/**
